Add cookie matching and split integration tests
