+++ mokksy/api/mokksy.klib.api
@@ -615,6 +615,8 @@ open class <#A: kotlin/Any> dev.mokksy.mokksy.request/RequestSpecification { // 
         final fun <get-body>(): kotlin.collections/List<io.kotest.matchers/Matcher<#A?>> // dev.mokksy.mokksy.request/RequestSpecification.body.<get-body>|<get-body>(){}[0]
     final val bodyString // dev.mokksy.mokksy.request/RequestSpecification.bodyString|{}bodyString[0]
         final fun <get-bodyString>(): kotlin.collections/List<io.kotest.matchers/Matcher<kotlin/String?>> // dev.mokksy.mokksy.request/RequestSpecification.bodyString.<get-bodyString>|<get-bodyString>(){}[0]
+    final val cookies // dev.mokksy.mokksy.request/RequestSpecification.cookies|{}cookies[0]
+        final fun <get-cookies>(): kotlin.collections/List<io.kotest.matchers/Matcher<io.ktor.server.request/RequestCookies>> // dev.mokksy.mokksy.request/RequestSpecification.cookies.<get-cookies>|<get-cookies>(){}[0]
     final val headers // dev.mokksy.mokksy.request/RequestSpecification.headers|{}headers[0]
         final fun <get-headers>(): kotlin.collections/List<io.kotest.matchers/Matcher<io.ktor.http/Headers>> // dev.mokksy.mokksy.request/RequestSpecification.headers.<get-headers>|<get-headers>(){}[0]
     final val method // dev.mokksy.mokksy.request/RequestSpecification.method|{}method[0]
@@ -632,6 +634,8 @@ open class <#A: kotlin/Any> dev.mokksy.mokksy.request/RequestSpecificationBuilde
         final fun <get-body>(): kotlin.collections/MutableList<io.kotest.matchers/Matcher<#A?>> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.body.<get-body>|<get-body>(){}[0]
     final val bodyString // dev.mokksy.mokksy.request/RequestSpecificationBuilder.bodyString|{}bodyString[0]
         final fun <get-bodyString>(): kotlin.collections/MutableList<io.kotest.matchers/Matcher<kotlin/String?>> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.bodyString.<get-bodyString>|<get-bodyString>(){}[0]
+    final val cookies // dev.mokksy.mokksy.request/RequestSpecificationBuilder.cookies|{}cookies[0]
+        final fun <get-cookies>(): kotlin.collections/MutableList<io.kotest.matchers/Matcher<io.ktor.server.request/RequestCookies>> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.cookies.<get-cookies>|<get-cookies>(){}[0]
     final val headers // dev.mokksy.mokksy.request/RequestSpecificationBuilder.headers|{}headers[0]
         final fun <get-headers>(): kotlin.collections/MutableList<io.kotest.matchers/Matcher<io.ktor.http/Headers>> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.headers.<get-headers>|<get-headers>(){}[0]
     final val requestType // dev.mokksy.mokksy.request/RequestSpecificationBuilder.requestType|{}requestType[0]
@@ -653,6 +657,9 @@ open class <#A: kotlin/Any> dev.mokksy.mokksy.request/RequestSpecificationBuilde
     final fun bodyMatchesPredicates(kotlin/Array<out kotlin/Function1<#A?, kotlin/Boolean>>...): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.bodyMatchesPredicates|bodyMatchesPredicates(kotlin.Array<out|kotlin.Function1<1:0?,kotlin.Boolean>>...){}[0]
     final fun bodyString(io.kotest.matchers/Matcher<kotlin/String?>): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.bodyString|bodyString(io.kotest.matchers.Matcher<kotlin.String?>){}[0]
     final fun containsHeader(kotlin/String, kotlin/String): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.containsHeader|containsHeader(kotlin.String;kotlin.String){}[0]
+    final fun cookie(kotlin/String, kotlin/Function1<kotlin/String?, kotlin/Boolean>): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.cookie|cookie(kotlin.String;kotlin.Function1<kotlin.String?,kotlin.Boolean>){}[0]
+    final fun cookie(kotlin/String, kotlin/String): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.cookie|cookie(kotlin.String;kotlin.String){}[0]
+    final fun cookieAbsent(kotlin/String): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.cookieAbsent|cookieAbsent(kotlin.String){}[0]
     final fun method(io.kotest.matchers/Matcher<io.ktor.http/HttpMethod>): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.method|method(io.kotest.matchers.Matcher<io.ktor.http.HttpMethod>){}[0]
     final fun path(io.kotest.matchers/Matcher<kotlin/String>): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.path|path(io.kotest.matchers.Matcher<kotlin.String>){}[0]
     final fun path(kotlin/String): dev.mokksy.mokksy.request/RequestSpecificationBuilder<#A> // dev.mokksy.mokksy.request/RequestSpecificationBuilder.path|path(kotlin.String){}[0]
